+++ secure/secure.py
@@ -1,16 +1,22 @@
+# Header information and recommendations are adapted from:
+# - OWASP Secure Headers Project (https://owasp.org/www-project-secure-headers/)
+# - MDN Web Docs (https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers#security)
+# - web.dev (https://web.dev)
+# - The World Wide Web Consortium (W3C) (https://www.w3.org)
+
 from typing import Any, Dict, List, Optional, Tuple, Union
 
 from .headers import (
-    CacheControl,
-    ContentSecurityPolicy,
-    PermissionsPolicy,
-    ReferrerPolicy,
-    ReportTo,
-    Server,
-    StrictTransportSecurity,
-    XContentTypeOptions,
-    XFrameOptions,
-    XXSSProtection,
+    server,
+    cache_control,
+    content_security_policy,
+    permissions_policy,
+    referrer_policy,
+    report_to,
+    strict_transport_security,
+    x_content_type_options,
+    x_frame_options,
+    x_xss_protection,
 )
 
 
@@ -32,16 +38,24 @@ class Secure:
 
     def __init__(
         self,
-        server: Optional[Server] = None,
-        hsts: Optional[StrictTransportSecurity] = StrictTransportSecurity(),
-        xfo: Optional[XFrameOptions] = XFrameOptions(),
-        xxp: Optional[XXSSProtection] = XXSSProtection(),
-        content: Optional[XContentTypeOptions] = XContentTypeOptions(),
-        csp: Optional[ContentSecurityPolicy] = None,
-        referrer: Optional[ReferrerPolicy] = ReferrerPolicy(),
-        cache: Optional[CacheControl] = CacheControl(),
-        permissions: Optional[PermissionsPolicy] = None,
-        report_to: Optional[ReportTo] = None,
+        server: Optional[server.Server] = None,
+        hsts: Optional[
+            strict_transport_security.StrictTransportSecurity
+        ] = strict_transport_security.StrictTransportSecurity(),
+        xfo: Optional[x_frame_options.XFrameOptions] = x_frame_options.XFrameOptions(),
+        xxp: Optional[
+            x_xss_protection.XXSSProtection
+        ] = x_xss_protection.XXSSProtection(),
+        content: Optional[
+            x_content_type_options.XContentTypeOptions
+        ] = x_content_type_options.XContentTypeOptions(),
+        csp: Optional[content_security_policy.ContentSecurityPolicy] = None,
+        referrer: Optional[
+            referrer_policy.ReferrerPolicy
+        ] = referrer_policy.ReferrerPolicy(),
+        cache: Optional[cache_control.CacheControl] = cache_control.CacheControl(),
+        permissions: Optional[permissions_policy.PermissionsPolicy] = None,
+        report_to: Optional[report_to.ReportTo] = None,
     ) -> None:
         self.server = server
         self.hsts = hsts
@@ -65,16 +79,16 @@ class Secure:
         self,
     ) -> List[
         Union[
-            CacheControl,
-            ContentSecurityPolicy,
-            PermissionsPolicy,
-            ReferrerPolicy,
-            ReportTo,
-            Server,
-            StrictTransportSecurity,
-            XContentTypeOptions,
-            XFrameOptions,
-            XXSSProtection,
+            cache_control.CacheControl,
+            content_security_policy.ContentSecurityPolicy,
+            permissions_policy.PermissionsPolicy,
+            referrer_policy.ReferrerPolicy,
+            report_to.ReportTo,
+            server.Server,
+            strict_transport_security.StrictTransportSecurity,
+            x_content_type_options.XContentTypeOptions,
+            x_frame_options.XFrameOptions,
+            x_xss_protection.XXSSProtection,
         ]
     ]:
         headers = [
